HAMA-572 - fixed unnecessary (un)boxing, old style loops, etc.
git-svn-id: https://svn.apache.org/repos/asf/hama/trunk@1442488 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/hama/bsp/TaskLogServlet.java
@@ -181,12 +181,12 @@ public class TaskLogServlet extends HttpServlet {
 
     String sLogOff = request.getParameter("start");
     if (sLogOff != null) {
-      start = Long.valueOf(sLogOff).longValue();
+      start = Long.valueOf(sLogOff);
     }
 
     String sLogEnd = request.getParameter("end");
     if (sLogEnd != null) {
-      end = Long.valueOf(sLogEnd).longValue();
+      end = Long.valueOf(sLogEnd);
     }
 
     String sPlainText = request.getParameter("plaintext");
